Split tests for query and collection
